spark-redshift connector: Dont use staging tables for overwrite operations
Summary:
This way we eliminate copying the data to a staging table and then inserting all of those rows into a prod table.
This behavior is needed for appends but on an overwrite where we are starting from scratch on a new table we can just copy directly to the new table.

Test Plan: Unit tests and etl test run

Reviewers: #investigation_data_services_team, nia.schmald

Reviewed By: #investigation_data_services_team, nia.schmald

Subscribers: nia.schmald

Differential Revision: https://signifyd.phacility.com/D19549